% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffEqModels.R
\name{SEIRS}
\alias{SEIRS}
\title{Generate an SEIRS model for an infection based on input parameters}
\usage{
SEIRS(
  infection,
  duration,
  time_step,
  N,
  percent_susceptible,
  number_infectious,
  infection_data
)
}
\arguments{
\item{infection}{The name of the infection to model c("influenza", "COVID").}

\item{duration}{The length of time to model, in days (optional; 120).}

\item{time_step}{The time between each model step, in days (optional; 1).}

\item{N}{The total population size (optional; 10^6).}

\item{percent_susceptible}{The initial susceptible population, as a percentage of N (optional; 100).}

\item{number_infectious}{The initial number of infectious individuals (optional; 1).}

\item{infection_data}{A list of key-value pairs with basic infection data. See simple_infections.}
}
\description{
Generate an SEIRS model for an infection based on input parameters
}
