% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffEqModels.R
\docType{data}
\name{simple_infections}
\alias{simple_infections}
\title{Some built-in data sets for common infections}
\format{
An object of class \code{data.frame} with 2 rows and 7 columns.
}
\usage{
simple_infections
}
\description{
Some built-in data sets for common infections
}
\keyword{datasets}
